requirements.txt
sentence-transformers
langchain
flask
aiofiles
fastapi
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain_experimental
langchain-huggingface

-e .